Floats to double

+++ PokeAgua.java
@@ -1,5 +1,5 @@
 public class PokeAgua extends Pokemon{
-		public PokeAgua (String apodo, float vida, float defensa, String estado, String tipo, float velocidad){
+		public PokeAgua (String apodo, double vida, double defensa, String estado, String tipo, double velocidad){
 		super(apodo,vida,defensa,estado,tipo,velocidad);
 	
 }
