+++ PokeFuego.java
@@ -1,5 +1,5 @@
 public class PokeFuego extends Pokemon {
-		public PokeFuego (String apodo, float vida, float defensa, String estado, String tipo, float velocidad){
+		public PokeFuego (String apodo, double vida, double defensa, String estado, String tipo, double velocidad){
 		super(apodo,vida,defensa,estado,tipo,velocidad);
 	
 }
