+++ PokeHierba.java
@@ -1,6 +1,6 @@
 public class PokeHierba extends Pokemon {
 	
-		public PokeHierba (String apodo, float vida, float defensa, String estado, String tipo, float velocidad){
+		public PokeHierba (String apodo, double vida, double defensa, String estado, String tipo, double velocidad){
 		super(apodo,vida,defensa,estado,tipo,velocidad);
 	
 }
